GitHub issue #11 - Removing chat messages when the action is completed by the bot.
Added a random button to the web UI.

GitHub issue #12 - Bot will play sound when user disconnects from a channel if there is a sound name "username_leave". The suffix is configurable in the app.properties file with the property "leave_suffix".

Per request on reddit, I've made it configurable if the bot responds to direct messages (private messages). This is configurable by a property in the app.properties file with the property "respond_to_dm". Default value is true.